+++ cmd/agent_system/mobilebind/agent.go
@@ -3,32 +3,11 @@ package mobilebind
 import (
 	"path"
 
-	log "github.com/sirupsen/logrus"
 	agentsystem "goauthentik.io/platform/pkg/agent_system"
-	systemlog "goauthentik.io/platform/pkg/platform/log"
 	"goauthentik.io/platform/pkg/platform/pstr"
 )
 
 var agent *agentsystem.SystemAgent
-
-func init() {
-	log.SetLevel(log.DebugLevel)
-	log.SetFormatter(&log.TextFormatter{
-		DisableTimestamp: true,
-		DisableColors:    true,
-	})
-}
-
-func InitSystemlog() bool {
-	err := systemlog.MustSetup(pstr.PlatformString{
-		Darwin: new("io.goauthentik.platform.app"),
-	}.ForCurrent())
-	if err != nil {
-		log.WithError(err).Warning("failed to setup system log")
-		return false
-	}
-	return true
-}
 
 func Init() bool {
 	a, err := agentsystem.New(agentsystem.SystemAgentOptions{
@@ -40,7 +19,7 @@ func Init() bool {
 		},
 	})
 	if err != nil {
-		log.WithError(err).Warning("failed to init agent")
+		logger.WithError(err).Warning("failed to init agent")
 		return false
 	}
 	agent = a
