+++ cmd/agent_system/mobilebind/facts.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 
-	log "github.com/sirupsen/logrus"
 	"goauthentik.io/api/v3"
 	"goauthentik.io/platform/pkg/platform/facts"
 	"goauthentik.io/platform/pkg/platform/facts/common"
@@ -12,13 +11,13 @@ import (
 )
 
 func Facts() string {
-	facts, err := facts.Gather(common.New(log.WithField("cmd", "facts"), context.Background()))
+	facts, err := facts.Gather(common.New(logger, context.Background()))
 	if err != nil {
-		log.WithError(err).Warning("failed to get all facts")
+		logger.WithError(err).Warning("failed to get all facts")
 	}
 	s, err := json.Marshal(facts)
 	if err != nil {
-		log.WithError(err).Warning("failed to json marshal")
+		logger.WithError(err).Warning("failed to json marshal")
 		return ""
 	}
 	return string(s)
